pyproject.toml
[tool.poetry]
name = "django-drf-polls"
version = "0.1.0"
description = "polls API"
authors = ["igor"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.1"
djangorestframework = "^3.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
